+++ packages/@overeng/oxc-config/nix/build.nix
@@ -14,7 +14,7 @@ import ../../../../nix/oxc-config-plugin.nix {
     ;
   # Managed by Evergreen FOD refresh — do not edit manually.
   depsBuilds = {
-    "." = mkSharedHash "sha256-8bxcNKrhmjaZch4rKGjmFqCbjvJFxwejfSQqi0bfecs=";
+    "." = mkSharedHash "sha256-Mi5gK1ZRBeCS8mbsPVIwXXmBGuDUyuar1T5WZbUMS+Y=";
   };
   hashSourcePath = "packages/@overeng/oxc-config/nix/build.nix";
 }